requirements.txt
requests>=2.31
pandas>=2.0
python-dotenv>=1.0
pyyaml>=6.0
plotly>=5.20
tqdm>=4.66
jupyter>=1.0
ipykernel>=6.29
dune-client>=1.7
